Added a new class: DataStore.
DataStore is a base class of DataStoreZabbix. In the future,
We might handle monitoring software other than ZABBIX.
So abstruct function should be gather in the basic class
such as this class.

A DataStore instance is created correspoding to the physical
monitoring server (such as Zabbix server).
VirtuaDataStore instance provides the composed data with
DataStore instances as if threre is only one monitoring server.

<< Next milestone >>
showing a 'Overview' of ZABBIX PHP UI. The assumed design of
the data stream is the following.

[Zabbix Server] =>
  (A) ArmZabbAPI -> (B) DataStoreZabbix -> (C) VirtualDataStoreZabbix
    -> (D) SQLProcessorZabbix -> (D) FaceMySQLWorker  => ZABBIX PHP client