Added a new class: DataStoreManager.
DataStoreManager manages DataStore instances. The instance of this class
is assumed to be held in a sub class of VirtualDataStore such as
VirtualDataStoreZabbix.

Because a sub class of VirtualDataStore should be used as a singleton,
The number of DataStoreManager instance is also assumed to be one.

<< Next milestone >>
showing a 'Overview' of ZABBIX PHP UI. The assumed design of
the data stream is the following.

[Zabbix Server] =>
  (A) ArmZabbAPI -> (B) DataStoreZabbix -> (C) VirtualDataStoreZabbix
    -> (D) SQLProcessorZabbix -> (D) FaceMySQLWorker  => ZABBIX PHP client