set DataStoreManager as a basic classic class of VirtualDataStore. Instead, DataStoreManager::m_dataStoreManager was deleted.
One of the side effects of this is able to remove
VirtualDataStore::add(). In addition, accessing DataStore
instances from VirtualDataStore becomes simple.

<< Tasks to do in the near future >>
- create DataStoreZabbix instance in mainRoutine().
- remove code to craete ArmZabbixAPI instance in ArmController.
  The ArmZabbixAPI instance will be created in DataStoreZabbix,
  When the above item is implemented.
- pass command line argument for ArmZabbiXAPI such as IP address etc.
  to DataSotreZabbix.

<< Next milestone >>
showing a 'Overview' (that is mainly needs triggers data of the server)
of ZABBIX PHP UI. The assumed design of the data stream is the following.

[Zabbix Server] =>
  (A) ArmZabbAPI -> (B) DataStoreZabbix -> (C) VirtualDataStoreZabbix
    -> (D) SQLProcessorZabbix -> (D) FaceMySQLWorker  => ZABBIX PHP client

<< Next next milestone >>
Merge data from two zabbix servers at (C) VirtualDataStoreZabbix