add functions to parse array and object combined JSON string to JsonParserAgent.
Added functions:
- bool JsonParserAgent::startElement(int index)
- void JsonParserAgent::endElement(void)

Now we can get a string in the array
without JsonParserAgent::read(int index, string &dest).
But we keep it for the convinience.

<< Tasks to do in the near future >>
- implement functions to read string and number value to JsonParserAgent
- implement DataStoreZabbix::getTriggers().
- implement VirtualDataStoreZabbix::getTriggers().
- parse arguments of --data-store-zabbix and create multiple
  DataStoreZabbix instances.

<< Next milestone >>
showing a 'Overview' (that is mainly needs triggers data of the server)
of ZABBIX PHP UI. The assumed design of the data stream is the following.

[Zabbix Server] =>
  (A) ArmZabbAPI -> (B) DataStoreZabbix -> (C) VirtualDataStoreZabbix
    -> (D) SQLProcessorZabbix -> (D) FaceMySQLWorker  => ZABBIX PHP client

<< Next next milestone >>
Merge data from two zabbix servers at (C) VirtualDataStore.