implemented the following function.
Implemented functions
- DataStoreZabbix::getTriggers().
- VirtualDataStoreZabbix::getTriggers().

This commit cannot still show 'Overview' of
ZABBIX PHP UI. Because it needs othre information
such as functions, items, and hosts.
We will add functions to get them.

And this patch just pass the ItemTablePtr
originally made in ArmZabbixAPI. It should be
merged in VirtualDatatoreZabbix in future commits.

<< Tasks to do in the near future >>
- implemented funtions to get the follwoing.
  + functions
  + items
  + hosts
- parse arguments of --data-store-zabbix and create multiple
  DataStoreZabbix instances.