fixed the problem that causes a crash in SQLProcessorSelect::makeTextRows().
The cause of the crash is invalid value of the active row.
Commit: 3d6adf2f56ebd4aa2c5271a23277826d00c9f2f9 changed the algorithm of
the join operation. In fact, it also affected the evaluation of formula in
makeTextRows(). Active rows are properly set only when SQLFromParser::doJoin()
is running. At the makeTextRows(), the active row points the outer of the table,
which doesn't has a meaningful value. (so SEGV might be caused)

To get correct data on makeTextRows(), we added a code to retrieve data
from SQLSelectInfo::selectedTables (or groupedTables), which are the
joined and satisfying the where condition.