added variables for indexing in ItemTable.h
The major motivation for adding an index is to improve performance
of inner join. The time to execute test_selectCountTrigger() is
about 86 [sec]. However, the time to execute the same SQL with
the MySQL client is 0.025 [sec]. So asura needs approx. 10^3 times
of machine resources.

I guess the reason is that current asura does inner join by
calculating all possible combinations of rows to check if the
condition matches. This may generate a lot of wasted processing.

We will implement the machanism that uses indexes to improve
performance (especially inner join) after this.