fixed the problem that makes test_test_groupByCountDistinct() fail.
The evaluation (calculation) of a formula of the column is performed
by every group (that is specified by 'GROUP BY' keyword). The older
version accumulates statistical value (that are for count(), max() and
so on) across all groups. As a result, it generates a wrong statistical
value.

The reason why FormulaElement::resetStatistics() is called in some
sub class's resetStatistics() is to call it in child formula elements.
However, code to call in chian has not been implemented yet.