added a new stub function: SQLProcessor::runForExists() as the first step to improve performance of procesing the statement including 'EXISTS'.
This patch is one of the preparations to improve performance for
processing the select statement including 'EXISTS'. The current
implementation calls select() to process the statement in 'EXISTS'_
per every condition check. This executes statemet parsing, making
tables, association of columns and so on many times. As a result,
the performance much degrades.