implemented ColumnComparisonPicker that is extract comparison between columns in where condition.
Currently the following patterns are supported
(a) A.B=C.D
(b) A.B=C.D AND E.F=G.H
(c) A.B=C.D AND E.F=G.H AND I.J=K.L AND ...

This patch only includes a test for the above (a). Tests for
other patterns will be added later.