added a virtual function: FormulaElement::removeParenthesis().
This method was introduced to remove FormulaParenthesis from the
formula tree. After the formula tree is built, its topology is not
changed. And a FormulaParenthesis instance just returns the child
value on evaluate(). Therefore, the instance in the tree has no
functional effect while it uses a little CPU resources.

There is also another motivation to introduce it. That is to make
finding primary conditions easy. Primary condition means a part
of formula that makes the entire formula FALSE when it is FALSE.

This patch hasn't included the implementation of removeParenthesis()
for FormulaParenthesis yet. It will be added later.