added a test for FaceRest (GET /servers).
Memo1:
test_servers() is a little long. We will extract some lines as
utility functions in the adding other tests.

Memo2:
The already developed test for Face class: testMySQLWorker
requires a running asura process. However, this test directly uses
FaceRest instance (i.e. doesn't need the asura process). I think
this method is more convenient and easier to setup libasura
than that of testMySQLWorker. So we want to fix testMySQLWorker
in this method.