added a concept of DBDomain, which enables to uses multiple database body.
The old version of DBAgent saves all tables in one database. However,
there are cases in which tables should be saved in separated databases.
So we added an argument for DBAgent: DBDomain ID. DBAgent instances
with the diffrent DBDomain ID uses the different database.
More details depend on the implementation of sub class of DBAgent.