implemented DBAgentSQLite3::select().
In commit: edae8702, we said 'DBAgentValue' would be used for SELECT.
However, this implementation doesn't uses DBAgentValue, but ItemData.
Because 'DBAgentValue' cannot make the copy of string, just has a
pointer of the string. This is not suitable for stroing string from
sqlite3_column_text().

In addition, I think that what there are silimilar structures is
undesirable. So we will remove 'DBAgentValue'.

After all the plan mentioned at commit: edae8702 was changed (cancelled).