added DBClientZabbix::resetDBInitializedFlags().
This function is reset all flags that indicate if the database is
initialized. Normally, it is only necessary to initialized once.
So Asura initializes the database at the start-up of the program,
memorizes it, and skips the initialization when the database is
opened after the second time.

However, tests delete and recreate database every test case. So
we need to invalidate information about the initialized databases.