added a new virtual method: DBAgent::deleteRows() and its implementation: DBAgentSQLite3::deleteRow().
This function performs the DELETE operation. Because the function
names for other SQL operations in DBAgent are 'select()', 'update()'
and 'insert()', the function name of this would be better to be
'delete()' simply. However, delete is a keyword of C++ and we
cannot use for the function name. So I added the suffix 'Rows'.