added a skipSetup flag to arguments of the DBAgent's constructor.
We will replace static methods that is used for initialization
(making tables and insert initial value etc.) in DBAgentSQLite3
with the normal DBAgent instances. Current implementation of
DBAgent and DBAgentSQLite3 suppose that the initialization is done
when it is created. And the instance automatically performs setup.
The setup naturally fails if the initialization has not been done.
So we have to skip the setup when we use a DBAgeint instance for
the initialization.