removed a multiple generation replica mechanism from replicated event table.
The generation replica mechanism keeps previoulsy obtained data.
This helps the investigation of server's critical bugs that lost
or break database. However, this fatures causes a flood of traffic
between servers and asura.

To replica monitoring data efficiently for the large monitored hosts
environment, we remove this feature and add a mechanism that get
data partially instead.