deleted a class ItemGroupTypeManager.
The class generates ItemGroupType instances based on array of
ItemData type. It caches the generated instance when it creates
first time. If the same ItemData array is specified, then the
ItemGroupTypeManager returns the cached ItemGroupType object.

This mechanism was introduced to save the memory and improve
performace. However, actually, it is difficult to delete the cache.
Since the cache is nerver deleted at present, it seems to be a
memory leak.

So we changed the way. The new idea is that the ItemGroup
instance simply has one ItemGroupType instance. And it is deleted
when the ItemGroup instance is destroyed.