moved a template method: ArmZabbixAPI::makeItemVector() to another file to make testing easy.
A template function is normally written in a header file.
However, I don't want to write makeItemVector() in ArmZabbixAPI.h
that can been seen from ArmZabbixAPI users, because makeItemVecotr()
is for the internal use.

A template function written in a source file cannot be instantiated
in other source file. This causes the difficulty of the test that
is written in another file. So I created special header file only
for the template function. This file assumed to be included only
in ArmZabbixAPI.cc and the test file.