modified the structure: EventInfo to store needed information directory.
Trigger conceptually expresses a condition, not a state.
So severity and status cannot be deferenced by the link to
trigger.

Becase each ZABBIX trigger has unitary severity value, the previous
data structure doesn't cause the problem. However, Naigos services
(that corresponds to trigger or item) can take several severity level.
So the data structure introduced in this patch is needed.