Check existence of mysql_version.h.
On RHEL6.4 or similars mysql_config is included in mysql package,
not in mysql-devel. So it's not enough only checking mysql_config,
we should also check header files.

Although it will be nice to replace checking mysql with AX_LIB_MYSQL
or something, I don't do it in this commit because it's out of scope
of this issue.