added the call of deleteDBClientDB(DB_DOMAIN_ID_HATOHOL) in setup() of testDBClientHatohol.
Previously only the test functions that need it call it in each test.
However, the number of such test functions is not small.
This fix makes the source code simple by removing deleteDBClientDB()
lines.

In addition, the function depends on the fact that the lower
layer of DBClientHatohol uses SQLite3. This fix is also intended
to reduce such dependency.