fixed implementation of DBClientHatohol::getNumberOfGoodHosts().
The previous version of getNumberOfGoodHosts() has a bug that
counts the number of hosts with bad trigger(s). The new approach
caluculates the difference between the total number of hosts and
the result of getNumberOfBadHosts().

As a side effect, this patch adds the getNumberOfHosts() that
returns the total number of hosts.