Replace the record ID management of DBClientAction with that based on the DB's auto increment function.
The current record ID management implementation of DBClientAction
may generate the ID that was used before after the record had been
deleted. This behaivor sometime causes troubles.
Instread we use that function based on the DB's feature that generates
monotonical increasing IDs.

As a side effect, we can make code of DBClientAction simple.