Add a stub of hatohol-resident-yard.
hatohol-resident-yard is a kind of container that executes
an action in it. The difference from the command-type action
is that hatohol-resident-yard keeps living after one action is
performed. This helps hatohol realize behaivor depending the
previous state.

For example, It is difficult for a command-type action to
pause the actions for a cetain time after one action is done,
because the command doesn't know the previous state.
Since hatohol-resident-yard keeps living, it can save previous
events in its memory and execute actions depending on it.