Change a structure for setup and connection inforation of DB.
I decided to clean up the code drastically, because
there are the following tree motivations.

(1) To remove a complicated setup mechanism
Previously, DB setup and connection information is stored in
the class: DBClientConnectable that is a template class and
has static members. In addtion, it manages connectionInfo
doubly. They are results of extension that was done every
time it were needed.

(2) Compatibility between compilers
clang compiler is incompatible with g++ in terms of the handle
of static member in a template class. They cause different result.
So we remove the template class itself that is one of the things
that make code complicated.

(3) To be the base class of DBClientZabbix
The previous structure doen't handle a DBClient derived classes
that uses a dynamically generated DB domain ID.