Replace the header file name.
DBClinetConnectable:
The previous base class that is a template class.

DBClinetConnectableBase:
The new base class that stores setup and connection information
in the map, instead of making some template bodies corresponding
to the DB domain ID.

This class name is a little long. I will change it latter or
merge it into DBClient.