Bump the version string and use singular form for functions.
This is the first commit for FaceRest protocol v2. The remaing
change will be commited later.

One of the major changes is the way to specify a format type of
the returned content. The new version uses a query for it.
Ex.)
Previous: /actions.jsonp/2354?callbackName=foo
New     : /action/2355?callbackName=foo&fmt=jsonp

I think the above style looks more like a REST API.