Remove dbAction from ActionManager::PrivateContext to improve safety.
ActionManager is used from some threads (mainly the ownere thread
that has an ActionManager instance, and GLib event loop thread).
Previously we restrict the use of PrivateContext::dbAction only
from the owner thread. However, now this restriction is very
complicated. So we remove it.

In this patch, a DBClientAction instance is passed as an argument
many times. This prevents the performance from being degraded due
to the execution of a DBClientActon's constructor that tries to
connect with a DB server.