Parse the siginfo on the ActorCollector thread.
Previously we use the asynchronus signal handler. So information about
the child is parsed on the GLIB's event handler where any function
can be called. Now we have the dedicated thread for waiting the child.
This patch makes the thread parse them to improve the performance by
removing communication between threads.