Make a DB Connection in the thread.
Previously ArnNagiosNDOUtils tries to connect with MySQL server
in the constructor. The constructor is called in the main thread
as a part of startup procedures. When it takes a while for the
connection, the subsequent startup procedures are also delayed.

This patch tries to connect in the startup of the thread.