[TravisCI] Install npm packages as a normal user.
On TravisCI, npm's global prefix is /home/travis/.nvm/v0.10.12
where the default user can install packages w/o a root permission.

On the other hand, when we execute 'npm' with sudo, the prefix is
/usr/local. As a result, the results of $ npm ls -g (in configure.ac)
and $ npm root -g (in Makefile.am) are different from /usr/local.
It causes the missing of installed packages.