[ActionManager] Add comments about the thread that call the function.
Methods in ActionManager are called from various threads. So we
have to recognize it.

Originally the motivation of this patch is to investigate of issue 67.
https://github.com/project-hatohol/hatohol/issues/67

I guess the problem is caused by the race between a function from
the GLIB event dispacher and deletion of the object including it
on other thread.