[server][SQLUtils] Add a new class: ItemDataCaster.
This class extract the actual ItemData instance (such as ItemInt
and ItemUint64) from a pointer of the base class: ItemData *
with SQLColumnType.
Since this class is implemented by the template, many things
is decided at a build time. As a result, overhead of calculating
the type to be casted can be small.