[client][users] Fix a bug that wrong users are deleted on deleting users.
For example, when an admin user try to delete a user with ID "12",
user "1" and user "2" are deleted in actual!
The reason is that the caller of HatoholItemRemover passes a string
"12" to it to specify items although it expects an array. As a
result, the string "12" is treated as an array "[1, 2]".