[server][ItemDataUtils] Overload << operator from ItemString to std::string.
Previous commits added a cast operator to string. However, it needs
an explicit cast to string in the following situation.

    string s;
    s = *itemData;         // compile error
    s = (string)*itemData; // OK

Because std::string can accept not only string but also int or bool
as a right hand side of '= operator', a compiler error happens
due to ambiguity of type to which *itemData is casted.

With this patch, the above code can be written as

    s << itemData;

This has two advantages. One is to omit the explicit cast. The other
is to omit '* operator'. As a result, the code becomes simpler.