[Server][FaceRest] Change function definition, name and signature.
* Remove a member function of FaceRest class.

* addHostgroupData ->
  buidHostgroup

Other functions aren't named such.
Because I think other functions name don't match a role of the function.
And, When I'll be free, I'll change those function name.

* agent ->
  outputJson

The reason is same as the above.
I think it is good that we will change function name and parameter name to match
a role of the function and parameter.