[Server][ArmZabbixAPI] Remove a function: updateHosts(const ItemTable *triggers)
Because, this function doesn't need.

Current mechanism of retrive host information get from only host id which is
included in triggers ItemTable.
But, I want to retrieve hostgroup id which is include in host.
An interval of current mechanism is very long.
(In some Cases, host and hostgroup id don't update.)