[server][DBClientHatohol] Filter triggers that belong to defunct servers.
NOTE: I am not satisfied with this implementation. It drops
triggers after the DBAgent gets them. DBAgent creates
ItemDataTable (including ItemGroup and ItemData instances)
and it is a little expensive (many memory allocations
are executed). So we should pass a filter condition to
an SQL statement.

I'll do it, right after this.