[server][SimpleSemaphore] Add a new class: SimpleSemaphore.
mlpl::EventSemaphore uses eventfd. So its post()/wait() always
enter kernel mode. However, it has a feature that can wait for
posts with poll().

On the other hand, SimpleSemaphore is based on the POSIX semaphore
whose implementation typically uses user-land code and futex.
So it's expected to be faster than EventSemaphroe in general.