[server] Add a way to append -std=c++11 for the C++ compiler.
By adding a option to the configure script like
$ ./configure --enable-cpp11
The compiler is executed with -std=c++11 option.

Now I won't use entire C++ features. I just want to use 'override'
keyword to avoid the degradation that might happen when change of
the parameter set of a virtual functions are leaked.

For example, when a paramter set of a base class's virtual method
is changed, the paramter set of all overridden methods must also be
changed. Or it won't work as a virtual method.

The present C++ specification doesn't have a mechanism to
detect this problem. However, such kind of the featur has been added
in C++11. It's a 'override' keyword. If this keyword is specified and
it doesn't actually override anything, the compiler causes an error.