[HatoholArmPluginInterface] Send initiation request soon after a plugin proecess starts.
For example, an passive arm-plugin process accidentally exits and
then an operator restarts it. The plugin process doesn't become ready state,
because initialization is not performed. Currently initialization packet
is sent from server (HatoholArmPluginGate) only when it starts.

With this patch, a plugin process requests to start initiation when it starts.