[server] When a user is deleted, remove the AccessInfo and ActionInfo associated with the user.
When remove the "User", to remove the "access_list" together, but don't remove "actions".
When displaying the "actions", to display only the "actions" that the user(owner) is exists.
To delete the "actions", run the deletion process on a cyclic(example:3600sec(1h)),
to remove an "actions" if there is no owner user.