[HatoholLoginDialog] Make browsers remember a user name the password.
A mechanism to remember values in an input form depends on browsers.
In general, they seem to rember a user name and the password when
a submit button is clicked in a form tag.

So this patch also makes an input form in the login dialog.

I confirmed that it works on Google Chrome 36.0 and Firefox 31.0.