[testActionManager] Fix a problem that fails test_checkEventsWithMultipleIncidentSender().
When HATOHOL_MYSQL_ENGINE_MEMORY=1 is set, the test fails.
The item order of DBAction:getActionList() depneds on some conditions
such as a memory engine and a expression of an SQL statement.

The patch calculates the expected action ID by calling getActionList()
actually with the similar condition of the Hatohol server.