[DBClientHatohol] Change the table name: servers to avoid a confilict.
Now there are the tables that have the same name.
One is owned by DBTablesConfig and the other is owned by
DBClientHatohol. They are currently different DB (MySQL and
SQLite3), it doesn't conflict.

However, we are fixing the architecuture. The final design will
have all tables in one DB in order to join each other.
In that case, the same table name cannot be allowed.