[Benchmark] Add an improve method to join strings.
The advantage of the join approach is to be able to get the total
length of the string that is finally crated. With the information,
the waste memory allocation can be reduced. As a result, performance
can be improved.

For example, the improve method (fast_join) was faster
than the SeparatorInjector.

*My envrionment
- Intel(R) Core(TM) i7 CPU         860  @ 2.80GHz
- Ubuntu 14.04

            Label:     Total   Average    Median
     string::join: (0.022s)  (0.002ms) (0.002ms)
string::fast_join: (0.006s)  (0.598us) (0.001ms)
SeparatorInjector: (0.008s)  (0.790us) (0.001ms)