[DBTablesMonitoring] Inherit DBTables.
Now DBTablesMonitoring is owned by DBHatohol, which manages a database connection.

The time of the unit test has increased by this change as below.

  Before previous: approx. 44 sec.
  With this patch: approx. 55 sec.

My environment:
  Core i7 i860 2.8GHz
  Ubuntu 14.04 64bit

Command:
  MLPL_LOGGER_LEVEL=WARN HATOHOL_MYSQL_ENGINE_MEMORY=1 ./run-test.sh --disable-signal-handling 2>/dev/null

The considered reason is that the performance of MySQL (memory storage
enginie) is slower than that of SQLite3. The monitoring data was stored
in the DB based on SQLite3 before. It is now stored in the DB based
on MySQL.

However, we should accept this performance degradation instead of
getting the following advantages.

- Be able to join with other tables. (It potentially improve performance)

- Easy to realize Act/Stanby system.
  In the previous Hatohol, a flood of copy from the monitoring server may occur.

- One of the step to realize Act/Act system.
  For example, it can be done with MariaDB/Galera replication.

NOTE: By this change, all the DBClient sub classes are replaced.