[IncidentSenderRedmine] Don't share a SoupSession among multiple IncidentSenderRedmine
Because it's possible to use different users against a same
Redmine server.

In addition, I expect that this approach will fix issue #290
(sometimes soup_auth_authenticate() crashes when it's called from
multiple threads simultaneously). When a SoupSession is separated,
SoupAuthManager is also separated and they will not use same
SoupAuth.