[client] base_ajax,hatohol_connector: Use POST mehtod for login to avoid exposing user name/password.
It is very bad to expose user's password in access logs.
To avoid that, let's use POST method for login.

Note that we need to get a token from django middleware
to cope with CSRF.  So far, form is used to get that token
as a cookie.  Later, we could use ensure_csrf_cookie().

Signed-off-by: YOSHIFUJI Hideaki <hideaki.yoshifuji@miraclelinux.com>