[HatoholServerEditDialogParameterized] Arrange JSON paramters.
This patch arrange paramter names and to improve readability and simplicity.

The new rules are following.
- Each paramter must have 'id' which is uniqueue.
- The paramter 'label' is used for the label field of UI.
- The WebUI use 'id' as key of returned parameters.

NOTE:
Currently REST service POST/PUT on '/server' saves conventional paramters only.
We will add a mechanism to save other paramters in the one column as a (probably JSON)
string, which is directly passed to Arm Plugins.