[hapi] Add monitoring of ArmPlugin(zabbix).
If an failuer in connection with the ArmPlugin or ArmPlugin in, then perform notice of event and register trigger.
Add a connection failure events of between the Hatohol(Arm) and Broker(qpid), and Broker(qpid) and ArmPlugin.
The events that occur in the ArmPlugin inside, notice from ArmPlugin to Hatohol during initialization.