[HatoholServerEditDialogParameterized] Show the dialog at the upper area of the window.
HatoholServerEditDialogParameterized fist shows select box and ADD/Cancel
buttons only at the center of the window. After a server type is selected,
parameters are appended and the dialog height is extended. However, the
top position of the dialog is not changed. As a result, the dialog shows
in the lower area. It sometimes make users scroll.

This patch avoid the inconvenience by showing put the initial dialog at
the upper area of the window.

I thought the another option that adjusts the dialog postion every
sever type change. However, I think moving the dialog makes users
tired or get confused.