[hapi] Add virtual functions.
onSetPluginInitialInfo : Called when starting the mainThread.
onFailureConnected : Called when connection with the AMQP broker is failure.
onMonitorArmPluginConnection : Called when started to wait for a AMQP broker's response.
onSuccessFetchMessage : Called when a AMQP broker's response is success.
onFailureFetchMessage : Called when a AMQP broker's response is failure.
onFailureReceivedMessage : Called when a HAPI's response is failure.