[ItemFetchWoker] Make a DataStore instance fetch items on demand.
Previously fetching items was in charged of ArmBase. However,
HatoholArmPluginGate is not a sub class of it. And I'm thinkg that
ArmBase should be removed (deprecated) because it was designed for
the conventional Arms such as ArmZabbixAPI and ArmNagiosNDOUtils.

So this patch suggests that a DataStore instance performs the task
instead of ArmBase.