[HapProcessCeilometer] Fix a problem in which old timestamp is sent.
Previously hap-process-ceilometer sends timestamp in GMT - TZ.
Since Hatohol server convert it to the local time, WebUI shows
the time as if it's in GMT.

In addition, the wrong old time is converted by localtime() when
the last event time is obtained. The use of localtime() is wrong
and doesn't make sense. But it unintentionally fixes the wrong time.

Now both send and get time are used in GMT.