Move a process to base_ajax.html from dashboard_ajax.html in a case to turn off
JavaScript of web browser. Currently, the process is applied only to hatohol
dashboard. But this state allows direct access to the other pages which hatohol
generates.

When this patch is applied, the process is applied to all pages. When hatohol
users direct access to all pages, this process redirects to disable_js.html.

FYI:
We announce to use the web browser with HTML5. If Hatohol users do not use HTML5,
we do not ensure correct behaviors. In general, 'noscript' tag in header is not
supported in HTML4.