mlpl ParsableString: Fix signedness issues which prevent us from using this module on multi-byte character strings.
An array (m_separatorArray[]) is used for checking if each character is
one of separators or not.  Array index was converted from char and
multi-byte characters are evaluated as negative value.  We screw up
here.

I do think we can assume that all separators must be evaluated as
positive intergers, but user-supplied string must not cause accessing
outside of the array.

Signed-off-by: YOSHIFUJI Hideaki <hideaki.yoshifuji@miraclelinux.com>